refactor: update parties to be slug-driven
