+++ src/routes/parties/[slug]/+page.ts
@@ -1,0 +1,16 @@
+import type { PageLoad } from './$types';
+import { error } from '@sveltejs/kit';
+import { partiesBySlug } from '$lib/data/all-parties';
+
+export const load: PageLoad = ({ params }) => {
+	const { slug } = params;
+	const party = partiesBySlug.get(slug);
+
+	if (!party) {
+		throw error(404, `Party "${slug}" not found`);
+	}
+
+	return {
+		party
+	};
+};
